Replace ITwitchTask with abstract base class This greatly reduces the amount of duplicate code
