+++ TwitchDownloaderWPF/TwitchTasks/TwitchTaskStatus.cs
@@ -1,0 +1,13 @@
+﻿namespace TwitchDownloaderWPF.TwitchTasks
+{
+    public enum TwitchTaskStatus
+    {
+        Waiting,
+        Ready,
+        Running,
+        Failed,
+        Finished,
+        Stopping,
+        Canceled
+    }
+}
